requirements.txt
urllib3
tornado
pyyaml

